Large code cleanup.
Move bitboard lookup tables into a separate file.
Move intrinsics into a separate file.
Use macros for FILE, RANK, SQUARE and SQUARE_BB in bitboard lookup generation.
Remove unused includes.

+++ src/tuning/fengen.cpp
@@ -162,7 +162,7 @@ void Tuning::fengen(std::string startPosPath, std::string outputPath, size_t num
                         break;
                     }
                     // If no mate is found and there are few enough pieces, it must be a draw
-                    else if(TB_LARGEST >= CNTSBITS(board.getColoredPieces(WHITE) | board.getColoredPieces(BLACK)))
+                    else if(TB_LARGEST >= (int) CNTSBITS(board.getColoredPieces(WHITE) | board.getColoredPieces(BLACK)))
                     {
                         result = DRAW;
                         break;
